refactor: move command handlers to commands module
Extract command handler functions from main.rs into dedicated files
under a new commands/ module for better code organization.

Closes #48
