+++ crates/cram/src/commands/mod.rs
@@ -1,0 +1,7 @@
+mod gui;
+mod list;
+mod self_update;
+
+pub use gui::launch_gui;
+pub use list::list_decks;
+pub use self_update::self_update;
